refactor(chart):Add chart for stations with sampling types.
