+++ src/components/SheetInfo.tsx
@@ -1,92 +1,11 @@
 import { useContext } from "react";
 import { Sheet, SheetContent, SheetDescription, SheetFooter, SheetHeader, SheetTitle } from "./ui/sheet";
-import type { MetObsIntervalValueType, MetObsSampleValueType } from "@/types/station";
-import { convertUnit, unit } from "@/utils/unit";
+import type { MetObsIntervalValueType } from "@/types/station";
+import { convertUnit } from "@/utils/unit";
 import { UnitContext } from "@/context/useUnitContext";
 import { SheetContext } from "@/context/useSheetContext";
 import { useStation } from "@/context/useStationContext";
-import { CartesianGrid, Line, LineChart, Tooltip, XAxis, YAxis, type TooltipContentProps } from 'recharts'
-import { useQuery } from "@tanstack/react-query";
-import { getStation } from "@/api/station";
-import { Parameter } from "@/context/useParameterContext";
-import type { ValueType, NameType } from "recharts/types/component/DefaultTooltipContent";
-
-function useStationValues(stationId: string) {
-    const parameterId = Parameter();
-
-    return useQuery({
-        queryKey: ["station", parameterId, stationId],
-        queryFn: () => getStation(parameterId, stationId),
-        enabled: !!stationId && !!parameterId,
-    })
-}
-
-/**
- * Converts the data to display on charts
- * @param {MetObsSampleValueType[]} values 
- * @returns 
- */
-function convertValues(values: MetObsSampleValueType[]) {
-    return values.map((value) => {
-        return {
-            ...value,
-            date: new Date(value.date),
-            value: parseInt(value.value),
-        }
-    })
-}
-
-const CustomTooltip = ({ active, payload, label }: TooltipContentProps<ValueType, NameType>) => {
-    const { unitType } = useContext(UnitContext);
-    const isVisible = active && payload && payload.length && unitType;
-    const date = new Date(label?.toString() as string).toLocaleString('sv-SE');
-    return (
-        <div className="custom-tooltip bg-white border-2 text-center p-2">
-            {isVisible && (
-                <>
-                    <p className="label">{`${date}`}</p>
-                    <p className="intro">{`${payload[0].value} ${unit[unitType]}`}</p>
-                </>
-            )}
-        </div>
-    );
-};
-
-
-function SamplingValuesTable({ stationId }: { stationId: string }) {
-    const { unitType } = useContext(UnitContext);
-    const { data, status } = useStationValues(stationId);
-
-    if (status === "error") return <div>Error</div>;
-    if (status === "pending") return <div>Loading...</div>;
-
-    if (!unitType) return <p>Undefined unit type</p>;
-
-    const values = convertValues(data.value as MetObsSampleValueType[]);
-
-    return (
-        <LineChart
-            width={500}
-            height={300}
-            data={values}
-            margin={{
-                top: 10,
-                right: 10,
-                bottom: 10,
-                left: 10
-            }}
-        >
-            <CartesianGrid strokeDasharray="3 3" />
-            <XAxis dataKey="date" tickFormatter={(string: Date) => {
-                const date = new Date(string).toLocaleString("sv-SE", { hour: '2-digit', minute: '2-digit' })
-                return date
-            }} />
-            <YAxis unit={unit[unitType]} />
-            <Tooltip content={CustomTooltip} />
-            <Line type="monotone" dataKey="value" stroke="#8884d8" activeDot={{ r: 8 }} />
-        </LineChart>
-    );
-}
+import StationSamplingChart from "./StationSamplingChart";
 
 function IntervalValuesTable({ values, }: { values?: MetObsIntervalValueType[] }) {
     const { unitType } = useContext(UnitContext);
@@ -124,7 +43,7 @@ export default function SheetInfo() {
                 </SheetHeader>
                 {
                     samplingValueType === "SAMPLING" ? (
-                        <SamplingValuesTable key={station.key} stationId={station.key} />
+                        <StationSamplingChart key={station.key} stationId={station.key} />
                     ) : (
                         <IntervalValuesTable values={station.value as MetObsIntervalValueType[]} />
                     )
