+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "strict": true,
     "esModuleInterop": true,
+    "noImplicitAny": true,
     "jsx": "react-jsx",
     "target": "ESNext",
     "module": "ESNext",
